pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "music-mcp-server"
version = "0.1.0"
description = "MCP server for controlling Spotify from GitHub Copilot Chat"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "John P" }
]
keywords = ["mcp", "spotify", "music", "copilot", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=0.1.0",
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
music-mcp-server = "music_mcp_server.server:main"
music-mcp-setup = "music_mcp_server.setup:main"

[project.urls]
Homepage = "https://github.com/jower999/music-mcp-server"
Repository = "https://github.com/jower999/music-mcp-server"
